pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dropbox>=12.0.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
]
